+++ src/code-snippets/component.angularjs.ts
