+++ src/code-snippets/service.angularjs.ts
