+++ src/commands/component.command.ts
@@ -1,5 +1,5 @@
 import { Command } from 'commander';
-import { generateComponent } from '../code-snippets/component-angularjs';
+import { generateComponent } from '../code-snippets/component.angularjs';
 import { VirtualFile } from '../models/virtual-file.model';
 import { createFile } from '../utilities/file-operations';
